requirements.txt
fastapi>=0.110,<1.0
uvicorn[standard]>=0.29,<1.0
python-dotenv
pandas>=2.2
numpy>=1.26
FinanceDataReader @ git+https://github.com/FinanceData/FinanceDataReader.git
pykrx
